fetch-football-togcs/requirements.txt
google-cloud-storage==2.18.0
google-cloud-pubsub==2.23.0
google-cloud-secret-manager==2.20.0
requests==2.32.3